refactor: extract settings + account handlers from TelegramService (PR 3/3)
Extract settings handlers into TelegramSettingsHandlers (telegram_settings.py)
and account handlers into TelegramAccountHandlers (telegram_accounts.py).

TelegramService reduced from ~1,984 to ~681 lines. The 3-PR refactor is now
complete, decomposing the original 3,504-line monolith into 6 focused modules.
