+++ CHANGELOG.md
@@ -8,6 +8,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+- **TelegramService Refactor PR 3: Extract Settings + Accounts** - Architecture improvement
+  - Extracted settings handlers into new `TelegramSettingsHandlers` class (`telegram_settings.py`)
+    - Handles: /settings command, toggle buttons, numeric edits (posts_per_day, hours), schedule management (regenerate/extend)
+  - Extracted account handlers into new `TelegramAccountHandlers` class (`telegram_accounts.py`)
+    - Handles: account selection menu, add/remove flows, inline account switching from posting workflow, back-to-post rebuild
+  - `TelegramService` reduced from ~1,984 to ~681 lines (core routing, initialization, captions, shared utilities)
+  - `_handle_callback` router updated to dispatch to `self.settings_handler.*` and `self.accounts.*`
+  - `_handle_conversation_message` delegates to new handler modules
+  - Tests split into `test_telegram_settings.py` and `test_telegram_accounts.py`
+  - All 345 tests pass (77 telegram-specific) - zero regressions
+  - Part 3 of 3 in the TelegramService decomposition plan (refactor complete)
+
 - **TelegramService Refactor PR 2: Extract Callbacks + Autopost** - Architecture improvement
   - Extracted 9 callback handlers into new `TelegramCallbackHandlers` class (`telegram_callbacks.py`)
     - Handles: posted, skipped, back, reject confirmation, cancel reject, rejected, resume, reset
